refactor : 이전 3단계 리뷰 피드백 반영
- schema 변경
- 누락사항 변경
- equals and hashcode 변

+++ src/main/java/nextstep/session/domain/StudentRepository.java
@@ -3,7 +3,7 @@ package nextstep.session.domain;
 import java.util.Optional;
 
 public interface StudentRepository {
-    int save(Student student);
+    Long save(Student student);
 
     Optional<Student> findById(Long studentId);
 }
